Add files via upload
The mobile application is built using flutter SDK and dart. Firebase is used for push notification and cloud messaging purposes. The application provides real-time location updates to the database to correlate with the data from the IoT device to get a useful output.
Application in detail:
This is a cross-platform mobile application. IT comes with a screen for login or sign up. Upon successful login, the home screen opens up with four buttons: status, ride hours, maps, and recommendations. It also has a rewarding scheme that is linked with amazon where the users could redeem their rewards.
The status page provides when the user was drowsy and how the user was detected drowsy.
The maps page leads to google maps which could be easy for the navigation for the driver (user).
The recommendations page provides the recommendations for the driver to be safe on the road.
The ride hours page provides how long the driver has been on the road. Which in turn helps the driver to get a better idea of how much he should get.
The app also uses acloud messaging service using firebase. So that the users could be sent timely notifications.